# perf(middleware): reduce chain allocation