Benchstat:
```
name                old time/op    new time/op    delta
ChainMiddlewares-4    2.69µs ± 2%    1.29µs ± 2%  -52.01%  (p=0.000 n=13+15)

name                old alloc/op   new alloc/op   delta
ChainMiddlewares-4    1.57kB ± 0%    0.48kB ± 0%  -69.39%  (p=0.000 n=15+15)

name                old allocs/op  new allocs/op  delta
ChainMiddlewares-4      39.0 ± 0%      20.0 ± 0%  -48.72%  (p=0.000 n=15+15)
```